requirements.txt
dj-database-url==0.5.0
Django==2.1
django-authtools==1.6.0
django-environ==0.4.5
django-heroku==0.3.1
gunicorn==19.9.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
whitenoise==4.0
